Upcast logits to fp32 in compute_masked_ce_loss to prevent bf16 NaN
The from-scratch SFT loss was calling F.cross_entropy directly on bf16
logits. With Qwen2.5-0.5B in bf16 (152k-token vocab), the softmax tail
probabilities underflow in bf16 -> log(0) = -inf -> loss = NaN within 20
training steps. Reproduced on RunPod A40 at batch=16, seq=1024.

Fix is the standard one used by HuggingFace, TRL, llama-recipes,
axolotl: .float() upcast on the logits just for the cross_entropy call.
Negligible memory/throughput cost (loss is a scalar).

Adds tests/test_loss_dtype.py covering: loss-finite with bf16 large-vocab
inputs, loss-is-fp32 even when logits are bf16, and a regression sanity
check on fp32 inputs.

+++ src/finpost/training/sft.py
@@ -56,8 +56,15 @@ def compute_masked_ce_loss(
     flat_logits = shift_logits.view(-1, shift_logits.size(-1))
     flat_labels = shift_labels.view(-1)
 
+    # Upcast logits to fp32 before cross_entropy. With a bf16 model
+    # (e.g. Qwen2.5-0.5B in bf16), `flat_logits` is bf16 and the softmax
+    # over a 150k+ token vocabulary underflows on the small tail
+    # probabilities -> log(0) = -inf -> loss = NaN. The fp32 upcast costs
+    # negligible memory/throughput (loss is a single scalar) and is the
+    # standard guard used by HuggingFace, TRL, llama-recipes, axolotl.
+    # See tests/test_loss_dtype.py for the regression test.
     return F.cross_entropy(
-        flat_logits,
+        flat_logits.float(),
         flat_labels,
         ignore_index=IGNORE_INDEX,
         reduction="mean",
